// apps/mobile/src/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      exercises: {
        Row: {
          author_id: string | null
          category: Database["public"]["Enums"]["exercise_category"]
          created_at: string
          id: string
          image_url: string | null
          info: string
          is_builtin: boolean
          is_public: boolean
          name: string
        }
        Insert: {
          author_id?: string | null
          category: Database["public"]["Enums"]["exercise_category"]
          created_at?: string
          id?: string
          image_url?: string | null
          info: string
          is_builtin?: boolean
          is_public?: boolean
          name: string
        }
        Update: {
          author_id?: string | null
          category?: Database["public"]["Enums"]["exercise_category"]
          created_at?: string
          id?: string
          image_url?: string | null
          info?: string
          is_builtin?: boolean
          is_public?: boolean
          name?: string
        }
        Relationships: [
          {
            foreignKeyName: "exercises_author_id_fkey"
            columns: ["author_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "exercises_author_id_fkey"
            columns: ["author_id"]
            isOneToOne: false
            referencedRelation: "profiles_public"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
    Enums: {
      exercise_category:
        | "hombros"
        | "pectorales"
        | "espalda"
        | "brazos"
        | "abdominales"
        | "piernas"
        | "estiramiento"
      user_type: "admin" | "gimnasio" | "entrenador" | "usuario" | "colaborador"
      weight_unit: "kg" | "lb" | "bloques"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never